rename fsnn modules
