+++ fsnn/vectors.py
